Added storybook and refactored redu store into store.js
