+++ ui-src/.storybook/addons.js
@@ -1,0 +1,2 @@
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';
