+++ ui-src/package.json
@@ -24,6 +24,15 @@
     "build": "react-scripts build && node copy-to-ui.js",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
-    "release:files": "node copy-release.js"
+    "release:files": "node copy-release.js",
+    "storybook": "start-storybook -p 9009 -s public",
+    "build-storybook": "build-storybook -s public"
+  },
+  "devDependencies": {
+    "@storybook/react": "^3.3.13",
+    "@storybook/addon-actions": "^3.3.13",
+    "@storybook/addon-links": "^3.3.13",
+    "@storybook/addons": "^3.3.13",
+    "babel-core": "^6.26.0"
   }
 }
