+++ ui-src/src/index.js
@@ -1,22 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { compact } from 'lodash'
-import { applyMiddleware, compose, createStore, combineReducers } from 'redux'
-import promiseMiddleware from 'redux-promise'
-import { requestSendingMiddleware, hcMiddleware } from 'hc-redux-middleware'
-import holochatReducer from './reducers'
 import Root from './root'
-import { reducer as formReducer } from 'redux-form'
-const middleware = compact([
-    hcMiddleware,
-    requestSendingMiddleware,
-    promiseMiddleware
-])
+import CreateStore from './store'
 
-const rootReducer = combineReducers({
-  holochatReducer,
-  form: formReducer
-})
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-let store = createStore(rootReducer, undefined, composeEnhancers(applyMiddleware(...middleware)))
+let store = CreateStore()
 ReactDOM.render(<Root store={store} />, document.querySelector('#root'))
